Файл main.js поделён на модули
