+++ js/store.js
@@ -1,0 +1,36 @@
+const store = {
+  _state: {
+    TYPE_OF_LIVING_ACCOMODATION: [
+      'palace',
+      'flat',
+      'house',
+      'bungalow',
+      'hotel',
+    ],
+    CHECK_IN_OUT_TIME: [
+      '12:00',
+      '13:00',
+      '14:00',
+    ],
+    FEATURES: [
+      'wifi',
+      'dishwasher',
+      'parking',
+      'washer',
+      'elevator',
+      'conditioner',
+    ],
+    PHOTOS: [
+      'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+      'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+      'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+    ],
+    COUNT_OF_ADVERTISEMENT: 10,
+  },
+  getState() {
+    return store._state; // Не могу использовать this... (\/)(O_o)(\/)
+  },
+};
+const getState = store.getState;
+
+export { getState };
